refactor: adjusting types, links and service
