+++ src/app/api/condominios/[condId]/users/[id]/route.ts
@@ -46,9 +46,9 @@ import { users } from '@/mocks/users';
  */
 export async function GET(
   _request: NextRequest,
-  { params }: { params: Promise<{ id: string }> }
+  { params }: { params: { id: string } }
 ) {
-  const { id } = await params;
+  const { id } = params;
   const user = users.find((u) => u.id === id);
 
   if (!user) {
@@ -97,9 +97,9 @@ export async function GET(
  */
 export async function PUT(
   request: NextRequest,
-  { params }: { params: Promise<{ id: string }> }
+  { params }: { params: { id: string } }
 ) {
-  const { id } = await params;
+  const { id } = params;
   const body = await request.json();
   console.log(`Updated information for user ${id}:`, body);
   return NextResponse.json(body);
@@ -144,9 +144,9 @@ export async function PUT(
  */
 export async function PATCH(
   request: NextRequest,
-  { params }: { params: Promise<{ id: string }> }
+  { params }: { params: { id: string } }
 ) {
-  const { id } = await params;
+  const { id } = params;
   const body = await request.json();
   console.log(`Patched information for user ${id}:`, body);
   return NextResponse.json(body);
@@ -178,9 +178,9 @@ export async function PATCH(
  */
 export async function DELETE(
   _request: NextRequest,
-  { params }: { params: Promise<{ id: string }> }
+  { params }: { params: { id: string } }
 ) {
-  const { id } = await params;
+  const { id } = params;
   console.log(`usuario com id ${id} foi apagado`);
   return NextResponse.json({ message: 'User deleted' });
 }
