+++ src/features/components/ui/sidebar.tsx
@@ -18,12 +18,12 @@ import {
   TooltipProvider,
   TooltipTrigger,
 } from '@/features/components/ui/tooltip';
+import { useIsMobile } from '@/features/hooks/use-mobile';
 import { Slot } from '@radix-ui/react-slot';
 import { cva, type VariantProps } from 'class-variance-authority';
 import { PanelLeft } from 'lucide-react';
 
 import { cn } from '@/lib/utils';
-import { useIsMobile } from '@/hooks/use-mobile';
 
 const SIDEBAR_COOKIE_NAME = 'sidebar_state';
 const SIDEBAR_COOKIE_MAX_AGE = 60 * 60 * 24 * 7;
